pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Shopify Voice AI assistant"
requires-python = ">=3.10, <3.15"

dependencies = [
    "livekit-agents>=1.5.1",
    "livekit-plugins-openai>=0.10.2",
    "livekit-plugins-silero>=0.6.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "livekit-plugins-deepgram>=1.5.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
